rename services and endpoints related to #671
