+++ src/server/pathway-commons/index.js
@@ -47,7 +47,7 @@ const PathwayCommonsService = {
     let url = 'https://www.pathwaycommons.org/pc2/get?' + qs.stringify(options);
     return fetch(url, fetchOptions).then(response => response.text());
   },
-  
+
   //Queries PC2 using traverse
   //Returns metadata for the pathway
   //Requires a valid query and uri
@@ -56,22 +56,22 @@ const PathwayCommonsService = {
     let url = 'https://www.pathwaycommons.org/pc2/traverse?' + qs.stringify(options);
     return fetch(url, fetchOptions).then(response => response.json());
   },
-  
+
   search(options) {
     //Construct query url
-    let url = 'https://www.pathwaycommons.org/pc2/search?' + qs.stringify(options); 
+    let url = 'https://www.pathwaycommons.org/pc2/search?' + qs.stringify(options);
     return fetch(url, fetchOptions).then(response => response.json());
   },
-  
+
   graph(options) {
     //Construct query url
-    let url = 'https://www.pathwaycommons.org/pc2/graph?' + qs.stringify(options); 
+    let url = 'https://www.pathwaycommons.org/pc2/graph?' + qs.stringify(options);
     return fetch(url, fetchOptions).then(response => response.json());
   },
-  
+
   top_pathways(options) {
     //Construct query url
-    let url = 'https://www.pathwaycommons.org/pc2/top_pathways?' + qs.stringify(options); 
+    let url = 'https://www.pathwaycommons.org/pc2/top_pathways?' + qs.stringify(options);
     return fetch(url, fetchOptions).then(response => response.json());
   }
 };
